+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # main
 
 * Added cop `RSpec/CreateListMax`
+* Added cop `Grape/HelpersIncludeModule` ([#1](https://github.com/petalmd/rubocop-petal/pull/1))
 
 # v0.1.0